refactor(dunning): migrate DeleteCampaignDialog to new dialog system
Replaces the legacy imperative ref-based WarningDialog with the hook-based
useCentralizedDialog from ~/components/dialogs, and relocates the file from
~/components/settings/dunnings to colocate with its parent under
~/pages/settings/Dunnings/dialogs.
